Add Follow/Unfollow functionality (added InteractionClient, updated on ServiceUser, AppSession and FollowInstance).

+++ src/main/java/org/hubson404/miniSocialNetwork/model/FollowInstance.java
@@ -15,7 +15,7 @@ public class FollowInstance {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int followId;
+    private Long followId;
 
     @ManyToOne
     @EqualsAndHashCode.Exclude
